+++ src/ui/views/DraftScouting/DraftClass.tsx
@@ -5,7 +5,7 @@ import {
 	LeagueFileUpload,
 	PlayerNameLabels,
 } from "../../components";
-import { downloadFile, getCols, toWorker } from "../../util";
+import { confirm, downloadFile, getCols, toWorker } from "../../util";
 import type { View } from "../../../common/types";
 
 const DraftClass = ({
@@ -84,6 +84,34 @@ const DraftClass = ({
 						}}
 					>
 						Regenerate
+					</button>
+				) : null}
+				{godMode ? (
+					<button
+						className="btn btn-god-mode btn-xs"
+						disabled={status === "exporting" || status === "loading"}
+						onClick={async () => {
+							setStatus("loading");
+
+							const proceed = await confirm(
+								`Are you sure you want to delete all ${players.length} players in the ${season} draft class?`,
+								{
+									okText: "Delete Players",
+								},
+							);
+
+							if (proceed) {
+								await toWorker(
+									"main",
+									"removePlayers",
+									players.map(p => p.pid),
+								);
+							}
+
+							setStatus(undefined);
+						}}
+					>
+						Clear
 					</button>
 				) : null}
 			</div>